Multiple minor improvements. Fix BEH_API_Remove test.
Use std::function rather than boost::signals for the transport callbacks, as we don't need the
multiple-slot facilities.  Added ManagedConnections::GetActiveConnectionsCount so that the tests can
identify when the connections are connected at both ends. Use [this] for lambda captures rather than
[&] as more explicit and safer. Incorporated Fraser's change to call the OnConnectionAdded callback
after the connection has been added to the internal table.