Fixes test ManagedConnectionsTest.BEH_API_Bootstrap
The test assumed there were no other instances of the
test running on the same PC and expected to fail
to connect to kLivePort. But when there was another
test program running, the RUDP node did sometimes manage to connect
to another node from the other test program, thus failing the
BEH_API_Bootstrap test.